movie_recommend_backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
mysql-connector-python==8.2.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.0
python-dotenv==1.0.0